Use Into rather than Deref
The code is less surprising and do not rely on knowing that KeyEvent
is wrapping a InputEvent as a new type generated by a macro, and that
Deref/AsRef return the InputEvent inside.

It seems to also be less confusing for LLM (who suggested to replace
*KeyEvent::new() by KeyEvent::new().clone() ), but I see that as a
feature that I remove (but there is likely others examples).