Add a warning about running as root
For some reason that I couldn't track down (yet), joycon keypresses
are not propagated to the process (but sync and absolute movement do).

While it could be a regression somewhere in the stack (linux, bluez), it
is also maybe a security feature, so I just want a warning until I get
down to it.